Stable alpha of double-post removal
After hours of trying to figure out why the reddit class wasn't returning the right post for the given id, discovered that reply_and_delete_submission was requesting the argument subm but was not using that argument for commenting and deletion, which locked it to whatever submission happened to be in the main loop.

Fixed that, returned a lot of test stuff back to how they were, and now have the script in a functional state for a temporary solution until it can be evaluated further.